pubsub-google-cloud/requirements.txt
google-api-python-client==2.58.0
google-auth-oauthlib==0.5.3
google-cloud-pubsub==2.21.1
fastapi